requirements.txt
gitpython==2.1.12
PyYAML==5.1.2
pylint==2.3.1
tqdm==4.33.0
future==0.17.1
